universal8895: Increase little cluster thermal cap
In addition to 2K+ resolution heavy gaming, to keep 60FPS as long as possible
with safe thermal settings, the frequencies needs to be capped at 1,7GHz and 1,4GHz
for little and big cores

My test game is Honkai Impact 3, which has graphics equal to PUBG mobile and
Fortnite. Setting the game at maximum graphics with 2K+ resolution resulted in
quick framerate drop due to little cluster capped at 1GHz, at a temperature of 66C

A simple solution for this is to slightly increase thermal cap of little cluster. Don't
raise the cap too high or the CPU will die as the transistors will "burn" (just physics
knowledge)

Signed-off-by: Diep Quynh <remilia.1505@gmail.com>