universal8895: Describe correct idle power costs
For 8895, we have no cluster and CPU off state, which leads to more power
consumption. This fact looks incorrect at all, and however, C2 state is
the maximum low power state that the CPU could reach

By the way, this is an unproven assumption so far. This commit will surely
get reverted in the future

Signed-off-by: Diep Quynh <remilia.1505@gmail.com>